+++ Plugins/AM_ProMex_Plugin/AnalysisToolRunnerProMex.cs
@@ -26,8 +26,8 @@ namespace AnalysisManagerProMexPlugIn
 
         private const string PROMEX_CONSOLE_OUTPUT = "ProMex_ConsoleOutput.txt";
 
-        private const float PROGRESS_PCT_STARTING = 1;
-        private const float PROGRESS_PCT_COMPLETE = 99;
+        private const int PROGRESS_PCT_STARTING = 1;
+        private const int PROGRESS_PCT_COMPLETE = 99;
 
         private string mConsoleOutputErrorMsg;
